ProyectoTienda/requirements.txt
psycopg2-binary==2.9.10
pycparser==2.22
crispy-bootstrap5==2025.4
Pillow==10.0.0
python-decouple
reportlab==3.6.12
openpyxl==3.1.2
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0